requirements.txt
gymnasium
matplotlib
numpy
pyyaml
tensorboardX
torch>=1.9.1
wandb>=0.17.4
setuptools
